implemented exceptions in Visit Notes' service layer

+++ src/main/java/com/markfox/patientmanager/services/impls/VisitNotesServiceImpl.java
@@ -1,5 +1,6 @@
 package com.markfox.patientmanager.services.impls;
 
+import com.markfox.patientmanager.exceptions.MyException;
 import com.markfox.patientmanager.models.VisitNotes;
 import com.markfox.patientmanager.repositories.VisitNotesRepository;
 import com.markfox.patientmanager.services.VisitNotesService;
@@ -16,12 +17,18 @@ public class VisitNotesServiceImpl implements VisitNotesService {
 
     // Implemented methods
     @Override
-    public VisitNotes addVisitNotes(VisitNotes notes) {
+    public VisitNotes addVisitNotes(VisitNotes notes) throws MyException {
+        if (notes == null) {
+            throw new MyException("Error: Visit Notes object cannot be null");
+        }
         return visitNotesRepository.save(notes);
     }
 
     @Override
-    public void deleteVisitNoteById(Long id) {
+    public void deleteVisitNoteById(Long id) throws MyException {
+        if (id == null) {
+            throw new MyException("Error: Visit Notes ID cannot be null");
+        }
         visitNotesRepository.deleteById(id);
     }
 }
